my-flask-app/app/requirements.txt
flask  
psycopg2-binary
